move hooks to root hook plugin
