+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "dev": "nodemon src/index.js",
     "start": "node src/index.js",
-    "test": "mocha src/test/*.test.js --timeout 10000 --exit"
+    "test": "mocha src/test/*.test.js --timeout 10000 --exit --require src/test/hooks.mjs"
   },
   "dependencies": {
     "axios": "^1.3.4",
