+++ src/test/tokens.validator.test.js
@@ -21,17 +21,6 @@ import { Database } from '../db/conn.js';
 chai.use(chaiHttp);
 
 describe('Tokens route - Validators', async function () {
-  /* The above code is a test cleanup function that runs after each test. It gets an instance of a
-database and checks if the namespace is 'grindery-delight-test-server'. If it is, it drops the
-'blockchains' collection from the database. This ensures that the database is cleaned up after each
-test and is ready for the next test. */
-  afterEach(async function () {
-    const db = await Database.getInstance({});
-    if (db.namespace === 'grindery-delight-test-server') {
-      db.dropDatabase();
-    }
-  });
-
   // Retry all tests in this suite up to 4 times
   this.retries(4);
